+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Changed
+- Add batch processing for bridge damage analysis [INCORE1-500](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-500)
+
 ## [0.6.0] - 2020-02-28
 
 ### Added